refactor: change structure components
