+++ class-components/src/components/Header/index.ts
@@ -1,1 +1,1 @@
-export { Header } from './Header.component';
+export { Header } from './Header';
