+++ class-components/src/components/Search/index.ts
@@ -1,0 +1,1 @@
+export { Search } from './Search';
